#68 Refactor: 테스트 코드의 skeleton 코드 패키지 구조 변경
