+++ TodaysLunch/src/test/java/LikeLion/TodaysLunch/member/controller/MemberControllerTest.java
@@ -1,5 +1,5 @@
 package LikeLion.TodaysLunch.member.controller;
 
-import LikeLion.TodaysLunch.skeleton.ControllerTest;
+import LikeLion.TodaysLunch.skeleton.controller.ControllerTest;
 class MemberControllerTest extends ControllerTest {
 }
