+++ TodaysLunch/src/test/java/LikeLion/TodaysLunch/member/service/MemberServiceTest.java
@@ -10,8 +10,8 @@ import LikeLion.TodaysLunch.exception.NotFoundException;
 import LikeLion.TodaysLunch.member.domain.Member;
 import LikeLion.TodaysLunch.member.dto.AdminJoinDto;
 import LikeLion.TodaysLunch.member.dto.MemberJoinDto;
-import LikeLion.TodaysLunch.skeleton.ServiceTest;
-import LikeLion.TodaysLunch.skeleton.TestUser;
+import LikeLion.TodaysLunch.skeleton.service.ServiceTest;
+import LikeLion.TodaysLunch.skeleton.service.TestUser;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
