+++ TodaysLunch/src/test/java/LikeLion/TodaysLunch/menu/service/MenuServiceTest.java
@@ -13,9 +13,9 @@ import LikeLion.TodaysLunch.menu.dto.MenuDto;
 import LikeLion.TodaysLunch.menu.dto.MenuImageDto;
 import LikeLion.TodaysLunch.menu.dto.SaleMenuDto;
 import LikeLion.TodaysLunch.menu.repository.MenuRepository;
-import LikeLion.TodaysLunch.skeleton.ServiceTest;
-import LikeLion.TodaysLunch.skeleton.TestRestaurant;
-import LikeLion.TodaysLunch.skeleton.TestUser;
+import LikeLion.TodaysLunch.skeleton.service.ServiceTest;
+import LikeLion.TodaysLunch.skeleton.service.TestRestaurant;
+import LikeLion.TodaysLunch.skeleton.service.TestUser;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
