+++ TodaysLunch/src/test/java/LikeLion/TodaysLunch/restaurant/service/RestaurantServiceTest.java
@@ -10,9 +10,9 @@ import LikeLion.TodaysLunch.exception.NotFoundException;
 import LikeLion.TodaysLunch.restaurant.dto.JudgeRestaurantListDto;
 import LikeLion.TodaysLunch.restaurant.dto.RestaurantListDto;
 import LikeLion.TodaysLunch.restaurant.dto.RestaurantRecommendDto;
-import LikeLion.TodaysLunch.skeleton.ServiceTest;
-import LikeLion.TodaysLunch.skeleton.TestRestaurant;
-import LikeLion.TodaysLunch.skeleton.TestUser;
+import LikeLion.TodaysLunch.skeleton.service.ServiceTest;
+import LikeLion.TodaysLunch.skeleton.service.TestRestaurant;
+import LikeLion.TodaysLunch.skeleton.service.TestUser;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
