+++ TodaysLunch/src/test/java/LikeLion/TodaysLunch/review/service/ReviewServiceTest.java
@@ -7,9 +7,9 @@ import LikeLion.TodaysLunch.restaurant.domain.Restaurant;
 import LikeLion.TodaysLunch.review.domain.Review;
 import LikeLion.TodaysLunch.review.dto.ReviewDto;
 import LikeLion.TodaysLunch.review.repository.ReviewRepository;
-import LikeLion.TodaysLunch.skeleton.ServiceTest;
-import LikeLion.TodaysLunch.skeleton.TestRestaurant;
-import LikeLion.TodaysLunch.skeleton.TestUser;
+import LikeLion.TodaysLunch.skeleton.service.ServiceTest;
+import LikeLion.TodaysLunch.skeleton.service.TestRestaurant;
+import LikeLion.TodaysLunch.skeleton.service.TestUser;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
