+++ TodaysLunch/src/test/java/LikeLion/TodaysLunch/skeleton/service/TestRestaurantEnviron.java
@@ -1,4 +1,4 @@
-package LikeLion.TodaysLunch.skeleton;
+package LikeLion.TodaysLunch.skeleton.service;
 
 import LikeLion.TodaysLunch.restaurant.repository.DataJpaRestaurantRepository;
 import LikeLion.TodaysLunch.menu.repository.MenuRepository;
