+++ TodaysLunch/src/test/java/LikeLion/TodaysLunch/skeleton/service/TestUserEnviron.java
@@ -1,4 +1,4 @@
-package LikeLion.TodaysLunch.skeleton;
+package LikeLion.TodaysLunch.skeleton.service;
 
 import LikeLion.TodaysLunch.customized.repository.MemberFoodCategoryRepository;
 import LikeLion.TodaysLunch.customized.repository.MemberLocationCategoryRepository;
